:sparkles: Add singleton for coredata accessor
